# docs/API_SPEC.yaml
openapi: 3.0.3
info:
  title: GGEE Backend API
  description: |
    브랜드 평판 방화벽 플랫폼 백엔드 API

    ## 인증
    - `/auth/**`, `/actuator/health`, `/actuator/info` 경로는 인증 없이 접근 가능
    - 그 외 모든 경로는 `Authorization: Bearer <accessToken>` 필요
    - 한 계정당 하나의 기기에서만 로그인 가능. 새 기기에서 로그인하면 이전 기기 세션이 무효화됨

    ## 토큰 만료
    - Access Token / Refresh Token 만료 시간은 환경 설정에 따라 달라질 수 있음
  version: 0.0.1

servers:
  - url: http://localhost:8080
    description: Local

tags:
  - name: Auth
    description: 회원가입 및 로그인
  - name: IndexingJob
    description: 인덱싱 작업 상태 조회

paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: 회원가입
      description: 새로운 사용자 계정을 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: "user@example.com"
              password: "password123"
              protectTarget: "brand"
      responses:
        "201":
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              example:
                indexing_job_id: "107519243230003200"
                accessToken: "eyJ0eXAiOiJqd3QiLCJhbGciOiJIUzI1NiJ9.xxxxx"
                refreshToken: "eyJ0eXAiOiJqd3QiLCJhbGciOiJIUzI1NiJ9.xxxxx"
        "400":
          description: 요청 본문 검증 실패
          content:
            text/plain:
              schema:
                type: string
              examples:
                blank_email:
                  value: "이메일이 입력되지 않았습니다."
                invalid_email:
                  value: "잘못된 이메일 형식입니다."
                blank_password:
                  value: "비밀번호가 입력되지 않았습니다."
                short_password:
                  value: "비밀번호는 최소 8자 이상 20자 이하여야 합니다."
                blank_protect_target:
                  value: "보호 대상을 입력해주세요."
        "409":
          description: 이미 존재하는 이메일
          content:
            text/plain:
              schema:
                type: string
              example: "이미 존재하는 이메일입니다."

  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: 로그인
      description: |
        이메일과 비밀번호로 로그인하여 JWT 토큰을 발급받습니다.
        로그인 시 기존 활성 세션이 모두 무효화됩니다. 이전 기기의 토큰은 더 이상 유효하지 않습니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "user@example.com"
              password: "password123"
      responses:
        "201":
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                accessToken: "eyJ0eXAiOiJqd3QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxMDc1MTkyNDMyMzAwMDMyIiwicm9sZSI6IlVTRVIiLCJ0b2tlblR5cGUiOiJBQ0NFU1NfVE9LRU4iLCJzZXNzaW9uSWQiOiJjYmQyZGU5ZC1mMmExLTQ2Y2EtOWJhZS0zZjIzNjk2YzY3ZGEiLCJpYXQiOjE3NzU0Mjg2ODksImV4cCI6MTc3NTQzMjI4OX0.xxxxx"
                refreshToken: "eyJ0eXAiOiJqd3QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxMDc1MTkyNDMyMzAwMDMyIiwicm9sZSI6IlVTRVIiLCJ0b2tlblR5cGUiOiJSRUZSRVNIX1RPS0VOIiwic2Vzc2lvbklkIjoiY2JkMmRlOWQtZjJhMS00NmNhLTliYWUtM2YyMzY5NmM2N2RhIiwiaWF0IjoxNzc1NDI4Njg5LCJleHAiOjE3NzY2MzgyODl9.xxxxx"
        "400":
          description: 요청 본문 검증 실패
          content:
            text/plain:
              schema:
                type: string
              examples:
                blank_email:
                  value: "이메일이 입력되지 않았습니다."
                invalid_email:
                  value: "잘못된 이메일 형식입니다."
                blank_password:
                  value: "비밀번호가 입력되지 않았습니다."
                short_password:
                  value: "비밀번호는 최소 8자 이상 20자 이하여야 합니다."
        "401":
          description: 이메일 또는 비밀번호 불일치
          content:
            text/plain:
              schema:
                type: string
              example: "이메일 또는 비밀번호가 올바르지 않습니다."

  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refresh
      summary: 토큰 갱신
      description: |
        Refresh Token으로 새로운 Access Token과 Refresh Token을 발급받습니다. (토큰 로테이션)
        Refresh Token이 만료된 경우 재로그인이 필요합니다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            example:
              refreshToken: "eyJ0eXAiOiJqd3QiLCJhbGciOiJIUzI1NiJ9.xxxxx"
      responses:
        "200":
          description: 토큰 갱신 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
              example:
                accessToken: "eyJ0eXAiOiJqd3QiLCJhbGciOiJIUzI1NiJ9.xxxxx"
                refreshToken: "eyJ0eXAiOiJqd3QiLCJhbGciOiJIUzI1NiJ9.xxxxx"
        "400":
          description: 요청 본문 검증 실패 또는 잘못된 토큰
          content:
            text/plain:
              schema:
                type: string
              examples:
                blank_token:
                  value: "refreshToken이 입력되지 않았습니다."
                invalid_token:
                  value: "올바르지 않은 JWT 토큰입니다."
                invalid_token_type:
                  value: "올바르지 않은 JWT 토큰 타입입니다."
                invalid_session:
                  value: "유효하지 않은 세션입니다."
        "401":
          description: 만료된 Refresh Token 또는 세션 만료
          content:
            text/plain:
              schema:
                type: string
              examples:
                expired_refresh_token:
                  value: "만료된 리프레시 토큰입니다."
                session_expired:
                  value: "다른 기기에서 로그인되어 세션이 만료되었습니다."

  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: 로그아웃
      description: |
        현재 인증된 사용자의 활성 세션을 무효화합니다.
        로그아웃 이후 기존 refresh token으로는 더 이상 토큰 갱신이 불가능합니다.
      security:
        - BearerAuth: []
      responses:
        "204":
          description: 로그아웃 성공
        "400":
          description: 잘못된 JWT 토큰
          content:
            text/plain:
              schema:
                type: string
              example: "올바르지 않은 JWT 토큰입니다."
        "401":
          description: 인증 실패 또는 세션 만료
          content:
            text/plain:
              schema:
                type: string
              examples:
                missing_token:
                  value: ""
                expired_token:
                  value: "만료된 JWT 토큰입니다."
                session_expired:
                  value: "다른 기기에서 로그인되어 세션이 만료되었습니다."

  /indexing/jobs/{jobId}:
    get:
      tags: [IndexingJob]
      operationId: streamJobCompletion
      summary: 인덱싱 작업 완료 대기 (SSE)
      description: |
        지정한 인덱싱 작업이 완료될 때까지 Server-Sent Events(SSE) 스트림으로 대기합니다.
        작업 완료 시 `completed` 이벤트가 전송됩니다.
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          description: 인덱싱 작업 ID (`/auth/register` 응답의 `indexing_job_id`)
          schema:
            type: string
          example: "107519243230003200"
      responses:
        "200":
          description: SSE 스트림 연결 성공. 작업 완료 시 completed 이벤트 수신
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                event: completed
                data: done
        "401":
          description: 인증 실패 (토큰 없음, 만료, 세션 만료 등)
          content:
            text/plain:
              schema:
                type: string
              examples:
                expired_token:
                  value: "만료된 JWT 토큰입니다."
                invalid_token:
                  value: "올바르지 않은 JWT 토큰입니다."
                session_expired:
                  value: "다른 기기에서 로그인되어 세션이 만료되었습니다."

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Authorization: Bearer <accessToken>"

  schemas:
    RegisterRequest:
      type: object
      required: [email, password, protectTarget]
      properties:
        email:
          type: string
          format: email
          description: 사용자 이메일
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 20
          description: 비밀번호 (8~20자)
          example: "password123"
        protectTarget:
          type: string
          description: 보호 대상 브랜드 또는 키워드
          example: "brand"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: 사용자 이메일
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 20
          description: 비밀번호 (8~20자)
          example: "password123"

    RegisterResponse:
      type: object
      properties:
        indexing_job_id:
          type: string
          description: "인덱싱 작업 ID. `/indexing/jobs/{jobId}` SSE 엔드포인트에서 완료 여부를 확인하는 데 사용"
          example: "107519243230003200"
        accessToken:
          type: string
          description: "JWT 액세스 토큰. Claims: sub(userId), role, tokenType, sessionId, iat, exp"
        refreshToken:
          type: string
          description: "JWT 리프레시 토큰. Claims: sub(userId), role, tokenType, sessionId, iat, exp"

    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: "JWT 액세스 토큰. 만료 시간은 환경 설정에 따름. Claims: sub(userId), role, tokenType, sessionId, iat, exp"
        refreshToken:
          type: string
          description: "JWT 리프레시 토큰. 만료 시간은 환경 설정에 따름. Claims: sub(userId), role, tokenType, sessionId, iat, exp"

    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: 로그인 또는 이전 갱신 시 발급받은 Refresh Token

    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: "새로 발급된 JWT 액세스 토큰"
        refreshToken:
          type: string
          description: "새로 발급된 JWT 리프레시 토큰 (토큰 로테이션)"